// src/lib/pe/types.ts
export interface SectionHeader {
  name: string;
  virtualSize: number;
  virtualAddress: number;
  sizeOfRawData: number;
  pointerToRawData: number;
  characteristics: number;
}

export interface DataDirectory {
  virtualAddress: number;
  size: number;
}

export interface COFFHeader {
  machine: number;
  numberOfSections: number;
  timeDateStamp: number;
  sizeOfOptionalHeader: number;
  characteristics: number;
}

export interface ResourceLanguage {
  languageId: number;
  dataRva: number;
  dataSize: number;
  codePage: number;
}

export interface ResourceEntry {
  id: number | null;
  name: string | null;
  languages: ResourceLanguage[];
}

export interface ResourceType {
  typeId: number | null;
  typeName: string | null;
  typeLabel: string;
  entries: ResourceEntry[];
}

export interface MZHeader {
  e_cblp: number;
  e_cp: number;
  e_crlc: number;
  e_cparhdr: number;
  e_minalloc: number;
  e_maxalloc: number;
  e_ss: number;
  e_sp: number;
  e_ip: number;
  e_cs: number;
  e_lfarlc: number;
  e_ovno: number;
}

export interface PEInfo {
  dosHeader: { e_magic: number; e_lfanew: number };
  coffHeader: COFFHeader;
  optionalHeader: { magic: number; isPE32Plus: boolean; subsystem: number; dataDirectories: DataDirectory[] };
  sections: SectionHeader[];
  resources: ResourceType[] | null;
  isNE?: boolean;
  neEncoding?: string;
  isMZ?: boolean;
  mzHeader?: MZHeader;
}
// ...
export interface DialogItem {
  style: number;
  exStyle: number;
  x: number;
  y: number;
  cx: number;
  cy: number;
  id: number;
  className: string;
  text: string;
  titleOrdinal: number | null;
}

export interface DialogFont {
  pointSize: number;
  weight?: number;
  italic?: boolean;
  typeface: string;
}

export interface DialogTemplate {
  style: number;
  exStyle: number;
  x: number;
  y: number;
  cx: number;
  cy: number;
  title: string;
  className: string | null;
  menuName: string | null;
  font: DialogFont | null;
  items: DialogItem[];
}

export interface DialogResult {
  id: number | null;
  name: string | null;
  languageId: number;
  dialog: DialogTemplate;
}